# docs/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "The Source API",
    "version": "1.0.0",
    "description": "Read-only Project Gutenberg catalog, installed text, and random prose. English is the default for selection endpoints. GET and HEAD only. No authentication. Query parameters must be supplied once. Browse and random endpoints reject unknown query parameters. Random responses use Cache-Control: no-store. HTML UI routes are not API endpoints. This document uses JSON syntax, which is valid YAML 1.2."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same server as this specification (default port 45068)"
    }
  ],
  "security": [],
  "paths": {
...
    "/api/v1/lyrics/excerpts/random": {
      "get": {
        "operationId": "randomStanza",
        "summary": "Get a random stanza from a random song",
        "description": "Tries up to eight random matching songs (honouring the optional language, tag and view-count filters). A stanza is a run of consecutive non-blank lines between blank lines; lines that are bracketed section markers such as [Chorus] are dropped, and a stanza must have at least two lines. Chooses uniformly among eligible stanzas within a song. Returns the source song (without its full lyrics body) and the stanza lines.",
        "parameters": [
          {"$ref": "#/components/parameters/songLanguage"},
          {"$ref": "#/components/parameters/tag"},
          {"$ref": "#/components/parameters/views_from"},
          {"$ref": "#/components/parameters/views_to"}
        ],
        "responses": {
          "200": {"description": "Stanza with source song metadata.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LyricsExcerpt"}}}, "headers": {"Cache-Control": {"schema": {"type": "string", "const": "no-store"}}}},
          "400": {"$ref": "#/components/responses/Invalid"},
          "404": {"$ref": "#/components/responses/LyricsMissing"},
          "405": {"$ref": "#/components/responses/Method"},
          "422": {"$ref": "#/components/responses/NoStanza"},
          "500": {"$ref": "#/components/responses/LyricsError"}
        }
      },
      "head": {
        "operationId": "randomStanzaHead",
        "summary": "Get a random stanza from a random song (headers only)",
        "description": "Same selection/validation as GET; no response body.",
        "parameters": [
          {"$ref": "#/components/parameters/songLanguage"},
          {"$ref": "#/components/parameters/tag"},
          {"$ref": "#/components/parameters/views_from"},
          {"$ref": "#/components/parameters/views_to"}
        ],
        "responses": {
          "200": {"description": "Stanza with source song metadata.", "headers": {"Cache-Control": {"schema": {"type": "string", "const": "no-store"}}}},
          "400": {"description": "400 invalid_query: invalid, unknown or repeated selection parameter."},
          "404": {"description": "404 no_matching_songs when no song matches; lyrics_unavailable when no corpus is installed."},
          "405": {"description": "405 method_not_allowed. Allow: GET, HEAD."},
          "422": {"description": "422 no_suitable_excerpt: the sampled songs yielded no qualifying stanza."},
          "500": {"description": "500 lyrics_error: the lyrics database could not be queried."}
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "songLanguage": {
        "name": "language",
        "in": "query",
        "description": "Optional language filter for songs. Unlike books, there is no English default: omission or all removes the filter. Trimmed and case-insensitive.",
        "schema": {"type": "string", "minLength": 1},
        "examples": {"english": {"value": "en"}, "all": {"value": "all"}}
      },
      "tag": {
        "name": "tag",
        "in": "query",
        "description": "Optional genre filter (e.g. rap, pop, rock, rb, country, misc). Trimmed and case-insensitive.",
        "schema": {"type": "string", "minLength": 1},
        "example": "rap"
      },
...
      "views_from": {
        "name": "views_from",
        "in": "query",
        "description": "Inclusive lower bound on view count. Omit or 0 for no lower bound. Must not exceed views_to.",
        "schema": {"type": "integer", "minimum": 0},
        "example": 1000
      },
      "views_to": {
        "name": "views_to",
        "in": "query",
        "description": "Inclusive upper bound on view count. Omit or 0 for no upper bound.",
        "schema": {"type": "integer", "minimum": 0},
        "example": 100000
      }
    },
    "schemas": {
...
      "Song": {
        "type": "object",
        "required": ["id", "title", "artist", "views"],
        "properties": {
          "id": {"type": "integer", "minimum": 1, "example": 10},
          "title": {"type": "string", "example": "Killa Cam"},
          "artist": {"type": "string", "example": "Cam'ron"},
          "tag": {"type": "string", "description": "Genre tag. Omitted when unknown.", "example": "rap"},
          "language": {"type": "string", "description": "Language code. Omitted when unknown.", "example": "en"},
          "year": {"type": "integer", "description": "Release year; only plausible values are kept, junk is omitted.", "example": 2004},
          "views": {"type": "integer", "minimum": 0, "description": "Genius view count; useful as a popularity signal."},
          "features": {"type": "array", "items": {"type": "string"}, "description": "Featured artists, parsed from the source array. Omitted when none."},
          "lyrics": {"type": "string", "description": "Lyrics body. Present only where explicitly documented; omitted from listings, item metadata, random and excerpt responses."}
        }
      },
...
      "LyricsExcerpt": {
        "type": "object",
        "required": ["song", "lines"],
        "properties": {
          "song": {"$ref": "#/components/schemas/Song"},
          "lines": {"type": "array", "minItems": 2, "items": {"type": "string"}, "description": "The lines of one stanza, section markers removed."}
        }
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "responses": {
      "Invalid": {
        "description": "400 invalid_query: invalid, unknown or repeated selection parameter; invalid range.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Method": {
        "description": "405 method_not_allowed. Allow: GET, HEAD.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "LyricsMissing": {
        "description": "404: song_not_found for unknown song ids; no_matching_songs for empty random selection; lyrics_unavailable when no lyrics corpus is installed.",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
      "LyricsError": {
        "description": "500 lyrics_error: the lyrics database could not be queried.",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
      "NoStanza": {
        "description": "422 no_suitable_excerpt: the sampled songs yielded no qualifying stanza. Does not mean the whole corpus lacks stanzas.",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      }
    }
  }
}